third_party/WebKit/LayoutTests/http/tests/credentialmanager/passwordcredential-basics.html

<!DOCTYPE html>
<title>Credential Manager: PasswordCredential basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/serviceworker/resources/interfaces.js"></script>
<script>
test(function() {
    var credential = new PasswordCredential('id', 'pencil', 'name', 'https://example.com/icon.png');

    verify_interface('PasswordCredential', credential, {
        formData: 'object',
        id: 'string',
        name: 'string',
        iconURL: 'string',
        password: 'string',
        type: 'string'
    });

    assert_true(credential.formData instanceof FormData);
    assert_equals(credential.id, 'id');
    assert_equals(credential.name, 'name');
    assert_equals(credential.iconURL, 'https://example.com/icon.png');
    assert_equals(credential.password, 'pencil');
    assert_equals(credential.type, 'password');
}, 'Interfaces and attributes of PasswordCredential');

test(function() {
    assert_throws(new SyntaxError(), function () {
        var credential = new PasswordCredential('id', 'pencil', 'name', '-');
    });
}, 'Construct a PasswordCredential with an invalid icon URL.');

test(function() {
    var credential = new PasswordCredential('id', 'pencil', 'name');

    assert_equals(credential.id, 'id');
    assert_equals(credential.name, 'name');
    assert_equals(credential.iconURL, '');
    assert_equals(credential.password, 'pencil');
    assert_equals(credential.type, 'password');

}, 'Construct a PasswordCredential with an empty icon URL.');

test(function() {
    var credential = new PasswordCredential('id', 'pencil');

    assert_equals(credential.id, 'id');
    assert_equals(credential.name, '');
    assert_equals(credential.iconURL, '');
    assert_equals(credential.password, 'pencil');
    assert_equals(credential.type, 'password');

}, 'Construct a PasswordCredential with an empty name and icon URL.');
</script>
